+++ src/main/java/com/cdc/coupons/entities/Coupon.java
@@ -11,6 +11,7 @@ import javax.validation.constraints.Future;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -28,7 +29,7 @@ public class Coupon {
     @Column(name = "discount", nullable = false)
     @Positive
     @NotNull
-    private Double discount;
+    private BigDecimal discount;
 
     @Column(name = "expiration_date", nullable = false)
     @NotNull
@@ -40,7 +41,7 @@ public class Coupon {
     }
 
     public Coupon(@NotBlank String code,
-                  @NotNull @Positive Double discount,
+                  @NotNull @Positive BigDecimal discount,
                   @NotNull @Future LocalDate expirationDate) {
         this.code = code;
         this.discount = discount;
